+++ hypervisor/Makefile
@@ -136,6 +136,7 @@ S_SRCS += arch/x86/idt.S
 C_SRCS += arch/x86/ioapic.c
 C_SRCS += arch/x86/lapic.c
 C_SRCS += arch/x86/cpu.c
+C_SRCS += arch/x86/cpu_caps.c
 C_SRCS += arch/x86/mmu.c
 C_SRCS += arch/x86/e820.c
 C_SRCS += arch/x86/pagetable.c
